+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 3.2.0
+* Added compound key support. Simply pass an array of attribute names to `meta.primaryKey`.
 * Updated `Renderer#formatValue` to accept an array of values, which internally calls `formatArray`.
 
 # 3.1.0